Completed
Added a character, a pipe puzzle, and a switch.

Camera movement is done with the mouse, character movement with WASD and SPACE to jump.

Click when next to and facing an interactable object to interact with it (no visual or audio feedback atm)

Pipes turn when interacted with, and the switch with print a debug message when the puzzle is completed correctly.